GHC.Generics.Lift

Description

Lift polymorphic typeclass over product types using Generics.

This module has two main purposes:

Write methods to lift methods of typeclasses to product types

Write default/automatic instances for your typeclasses so users can auto-derive them

These functions can be used to lift polymorphic functions on any type (deriving Generic) that has a single constructor, where every field is an instance of the class constraint of the function being lifted.

See README for details on usage instructions and motivations.

Documentation

class GLift c f where

Minimal complete definition

glift0, glift1, glift2, glift3

Methods

glift0 :: (forall a. c a => a) -> f p

glift1 :: (forall a. c a => a -> a) -> f p -> f p

glift2 :: (forall a. c a => a -> a -> a) -> f p -> f p -> f p

glift3 :: (forall a. c a => a -> a -> a -> a) -> f p -> f p -> f p -> f p

genericLift0 :: forall a c. (Generic a, GLift c (Rep a)) => (forall b. c b => b) -> a

genericLift1 :: forall a c. (Generic a, GLift c (Rep a)) => (forall b. c b => b -> b) -> a -> a

genericLift2 :: forall a c. (Generic a, GLift c (Rep a)) => (forall b. c b => b -> b -> b) -> a -> a -> a

genericLift3 :: forall a c. (Generic a, GLift c (Rep a)) => (forall b. c b => b -> b -> b -> b) -> a -> a -> a -> a