Numeric.Generics

Description

Derived methods for numeric typeclasses.

Can be used for any types (deriving Generic) made with a single constructor, where every field is an instance of Num (or Fractional or Floating, depending on the function).

Also includes a newtype wrapper that imbues any such data type with an instant Num (and Fractional and Floating) instance.

See README for details on usage instructions and motivations.

Newtype wrapper

newtype GNum a

If a is a data type with a single constructor whose fields are all instances of Num, then GNum a has a Num instance.

If a is a data type with a single constructor whose fields are all instances of Fractional, then GNum a has a Fractional instance.

If a is a data type with a single constructor whose fields are all instances of Floating, then GNum a has a Floating instance.

Constructors

GNum

Fields

getGNum :: a

Generics-derived methods

Num

genericPlus :: forall a. (Generic a, GLift Num (Rep a)) => a -> a -> a

genericMinus :: forall a. (Generic a, GLift Num (Rep a)) => a -> a -> a

genericTimes :: forall a. (Generic a, GLift Num (Rep a)) => a -> a -> a

genericNegate :: forall a. (Generic a, GLift Num (Rep a)) => a -> a

genericAbs :: forall a. (Generic a, GLift Num (Rep a)) => a -> a

genericSignum :: forall a. (Generic a, GLift Num (Rep a)) => a -> a

genericFromInteger :: forall a. (Generic a, GLift Num (Rep a)) => Integer -> a

Fractional

genericDivide :: forall a. (Generic a, GLift Fractional (Rep a)) => a -> a -> a

genericRecip :: forall a. (Generic a, GLift Fractional (Rep a)) => a -> a

genericFromRational :: forall a. (Generic a, GLift Fractional (Rep a)) => Rational -> a

Floating

genericPi :: forall a. (Generic a, GLift Floating (Rep a)) => a

genericExp :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericLog :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericSqrt :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericPower :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a -> a

genericLogBase :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a -> a

genericSin :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericCos :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericTan :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericAsin :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericAcos :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericAtan :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericSinh :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericCosh :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericTanh :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericAsinh :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericAcosh :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a

genericAtanh :: forall a. (Generic a, GLift Floating (Rep a)) => a -> a